requirements.txt
# 数据库驱动和连接池
clickhouse-driver>=0.2.0
clickhouse-pool>=0.1.0
clickhouse-cityhash>=1.0.2
redis>=4.0.0

# 数据处理
pandas>=1.3.0
numpy>=1.21.0

# 回测框架
backtrader>=1.9.76

# 数据可视化
matplotlib>=3.4.0

# 配置和工具
pyyaml>=5.4.0
python-dotenv>=0.19.0

# 测试工具
pytest>=6.2.5
pytest-cov>=2.12.0

# 类型提示
typing-extensions>=3.10.0

# 日志
loguru>=0.5.3

# 在数据库驱动部分添加
DBUtils>=3.0.0 